+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add session ID, transport, username, and IP addresses to log output.
 * Log number of TURN-relayed bytes and packets.
 * Log plain STUN (Binding) responses.
+* Log all error responses.
 
 # Version 1.0.34